[edit] type 선언 위치 수정 및 분리 적용
- types => interface 폴더로 이동
- 각 파일에 선언된 type 응집도 고려하여 위치 변경
