+++ src/interface/user.ts
@@ -1,0 +1,10 @@
+export type UserSelect = {
+  sizeId: number;
+  sizeValue: number;
+  ingredientIds: number[];
+};
+
+export type UserItem = {
+  sizeId: number;
+  ingredientIds: number[];
+};
